Split exceptions and fix tests
